Fix some more implicit double to float and double to int conversion
git-svn-id: http://libroxml.googlecode.com/svn/trunk@176 c40beb28-d10c-11dd-a2b5-e5f827957e07

+++ inc/roxml-internal.h
@@ -175,7 +175,7 @@ void ROXML_INT roxml_free_xpath			(xpath_node_t *xpath, int nb);
  * \param op the operator to use
  * \return 1 if comparison is ok, esle 0
  */
-double ROXML_INT roxml_double_cmp			(double a, double b, int op); 
+int ROXML_INT roxml_double_cmp			(double a, double b, int op); 
 
 /** \brief double operation function
  *
